pyproject.toml
[tool.poetry]
name = "signals"
version = "0.1.0"
description = ""
authors = ["Professional <pyostr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
klines = {git = "https://github.com/pyostr-traiding/klines.git"}
redis = "^6.4.0"
python-dotenv = "^1.1.1"
cachetools = "^6.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
